requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
celery>=5.3.0
redis>=5.0.0
pandas>=2.2.0
email-validator>=2.0.0
python-dotenv>=1.0.0
aiofiles==23.2.1
python-amazon-sp-api==0.19.0
requests==2.31.0
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0 